+++ src/main.cpp
@@ -20,15 +20,15 @@ int main()
 {
 
 	// parameters
-	const vector3<int64_t> volSize(600, 500, 300);
+	const vector3<int> volSize(600, 500, 300);
 	const float clipLimit = 0.1;
-	const int64_t binSize = 250;
-	const vector3<int64_t> subVolSize(31, 31, 31);
-	const vector3<int64_t> subVolSpacing(20, 20, 20);
+	const int binSize = 250;
+	const vector3<int> subVolSize(31, 31, 31);
+	const vector3<int> subVolSpacing(20, 20, 20);
 
 	// generate input volume matrix and assign random values to it
 	float* inputVol = new float[volSize.x * volSize.y * volSize.z];
-	for(int64_t iIdx = 0; iIdx < (volSize.x * volSize.y * volSize.z); iIdx ++)
+	for(int iIdx = 0; iIdx < (volSize.x * volSize.y * volSize.z); iIdx ++)
 		inputVol[iIdx] = ((float) rand()) / ((float) RAND_MAX);
 		// this should generate a random number between 0 and 1
