+++ src/singleFile_GPU.cu
@@ -22,21 +22,21 @@ using namespace std;
 // arguments structure passed to equilization kernel 
 struct eq_arguments
 {
-	int64_t volSize[3]; // total size of volume
+	int volSize[3]; // total size of volume
 	float origin[3]; // origin of the subvolume grid
 	float end[3]; // end of the subvolume grid
-	int64_t nSubVols[3]; // number of subvolumes
-	int64_t spacingSubVols[3]; // spacing between subvolumes
+	int nSubVols[3]; // number of subvolumes
+	int spacingSubVols[3]; // spacing between subvolumes
 	float* minValBin; // minimum value in each bib
 	float* maxValBin; // maximum value in each bin
 	float* cdf; // cummulative distribution function
-	int64_t nBins; // number of bins we have for the histogram
+	int nBins; // number of bins we have for the histogram
 };
 
 // returns the indices of the neighbouring subvolumes for a defined position
 __device__ inline void get_neighbours_gpu(
-	const int64_t* position,
-	int64_t* neighbours,
+	const int* position,
+	int* neighbours,
 	float* ratio,
 	const eq_arguments inArgs
 	)
@@ -60,7 +60,7 @@ __device__ inline void get_neighbours_gpu(
 		else // we are actually in between!
 		{
 			const float offsetDistance = ((float) position[iDim]) - (float) inArgs.origin[iDim];
-			neighbours[iDim * 2] = (int64_t) (offsetDistance / inArgs.spacingSubVols[iDim]);
+			neighbours[iDim * 2] = (int) (offsetDistance / inArgs.spacingSubVols[iDim]);
 			neighbours[iDim * 2 + 1] = neighbours[iDim * 2] + 1;
 			const float leftDistance = offsetDistance - ((float) neighbours[iDim * 2]) * 
 				inArgs.spacingSubVols[iDim];
@@ -73,14 +73,14 @@ __device__ inline void get_neighbours_gpu(
 
 // return inverted cumulative distribution function values
 __device__ inline float get_icdf_gpu(
-	const int64_t iZ, // index of subvolume we request along z
-	const int64_t iX, // index of subvolume we request along x
-	const int64_t iY, // index of subvolume we request along y
+	const int iZ, // index of subvolume we request along z
+	const int iX, // index of subvolume we request along x
+	const int iY, // index of subvolume we request along y
 	const float currValue,
 	const eq_arguments inArgs)
 {
 	// if we are below noise level, directy return 0
-	const int64_t subVolIdx = iZ + inArgs.nSubVols[0] * (iX + inArgs.nSubVols[1] * iY);
+	const int subVolIdx = iZ + inArgs.nSubVols[0] * (iX + inArgs.nSubVols[1] * iY);
 	if (currValue <= inArgs.minValBin[subVolIdx])
 	{
 		return 0;
@@ -88,13 +88,13 @@ __device__ inline float get_icdf_gpu(
 	else
 	{
 		// get index describes the 3d index of the subvolume
-		const int64_t subVolOffset = inArgs.nBins * subVolIdx;
+		const int subVolOffset = inArgs.nBins * subVolIdx;
 		const float vInterp = (currValue - inArgs.minValBin[subVolIdx]) / 
 			(inArgs.maxValBin[subVolIdx] - inArgs.minValBin[subVolIdx]); // should now span 0 to 1 
 		
 		// it can happen that the voxel value is higher then the max value detected
 		// in the next volume. In this case we crop it to the maximum permittable value
-		const int64_t binOffset = (vInterp > 1) ? 
+		const int binOffset = (vInterp > 1) ? 
 			(inArgs.nBins - 1 + subVolOffset)
 			: (vInterp * ((float) inArgs.nBins - 1.0) + 0.5) + subVolOffset;
 
@@ -119,7 +119,7 @@ __global__ void equalize_kernel(
 	)
 {
 	// get index of currently adjusted voxel
-	const int64_t idxVol[3] = {
+	const int idxVol[3] = {
 		threadIdx.x + blockIdx.x * blockDim.x,
 		threadIdx.y + blockIdx.y * blockDim.y,
 		threadIdx.z + blockIdx.z * blockDim.z
@@ -130,12 +130,12 @@ __global__ void equalize_kernel(
 		(idxVol[1] < inArgs.volSize[1]) && 
 		(idxVol[2] < inArgs.volSize[2]))
 	{
-		const int64_t idxVolLin = idxVol[0] + inArgs.volSize[0] * 
+		const int idxVolLin = idxVol[0] + inArgs.volSize[0] * 
 			(idxVol[1] + inArgs.volSize[1] * idxVol[2]);
 		const float currValue = dataMatrix[idxVolLin];
 
 		// get neighbours defined as the subvolume indices at lower and upper end
-		int64_t neighbours[6];
+		int neighbours[6];
 		float ratio[3];
 		get_neighbours_gpu(idxVol, neighbours, ratio, inArgs);
 
@@ -165,26 +165,26 @@ __global__ void equalize_kernel(
 // struct holding arguments used in cdf kernel
 struct cdf_arguments
 {
-	int64_t spacingSubVols[3]; // distance between subvolumes [z, x, y]
-	int64_t nSubVols[3]; // number of subvolumes [z, x, y]
-	int64_t volSize[3]; // overall size of data volume [z, x, y]
-	int64_t range[3]; // range of each bin in each direction [z, x, y]
-	int64_t nBins; // number of bins which we use for our histogram
+	int spacingSubVols[3]; // distance between subvolumes [z, x, y]
+	int nSubVols[3]; // number of subvolumes [z, x, y]
+	int volSize[3]; // overall size of data volume [z, x, y]
+	int range[3]; // range of each bin in each direction [z, x, y]
+	int nBins; // number of bins which we use for our histogram
 	float noiseLevel; // noise level in matrix
 	float origin[3];
 };
 
 // get start index limited by 0
-__device__ inline int64_t get_startIndex(const int64_t zCenter, const int zRange)
+__device__ inline int get_startIndex(const int zCenter, const int zRange)
 {
-	const int64_t startIdx = (((int) zCenter - zRange) < 0) ? 0 : zCenter - zRange;
+	const int startIdx = (((int) zCenter - zRange) < 0) ? 0 : zCenter - zRange;
 	return startIdx;
 }
 
 // get stop index limited by volume size
-__device__ inline int64_t get_stopIndex(const int64_t zCenter, const int zRange, const int64_t volSize)
+__device__ inline int get_stopIndex(const int zCenter, const int zRange, const int volSize)
 {
-	const int64_t stopIdx = (((int) zCenter + zRange) >= volSize) ? volSize : zCenter + zRange;
+	const int stopIdx = (((int) zCenter + zRange) >= volSize) ? volSize : zCenter + zRange;
 	return stopIdx;
 }
 
@@ -197,7 +197,7 @@ __global__ void cdf_kernel(
 		const cdf_arguments inArgs // constant arguments such as matrix size
 	)
 {
-	const int64_t iSub[3] = { // get index of current subvolume
+	const int iSub[3] = { // get index of current subvolume
 		threadIdx.x + blockIdx.x * blockDim.x,
 		threadIdx.y + blockIdx.y * blockDim.y,
 		threadIdx.z + blockIdx.z * blockDim.z
@@ -209,20 +209,20 @@ __global__ void cdf_kernel(
 		(iSub[2] < inArgs.nSubVols[2]))
 	{
 		// get start and stop indices for currently used bin
-		int64_t startIdx[3];	int64_t stopIdx[3];
+		int startIdx[3];	int stopIdx[3];
 		#pragma unroll
 		for (uint8_t iDim = 0; iDim < 3; iDim++)
 		{
-			const int64_t ctr = iSub[iDim] * inArgs.spacingSubVols[iDim] + inArgs.origin[iDim];
+			const int ctr = iSub[iDim] * inArgs.spacingSubVols[iDim] + inArgs.origin[iDim];
 			startIdx[iDim] = get_startIndex(ctr, inArgs.range[iDim]);
 			stopIdx[iDim] = get_stopIndex(ctr, inArgs.range[iDim], inArgs.volSize[iDim]);
 		}
 		
 		// index of currently calculated subvolume
-		const int64_t idxSubVol = iSub[0] + inArgs.nSubVols[0] * (iSub[1] + inArgs.nSubVols[1] * iSub[2]);
+		const int idxSubVol = iSub[0] + inArgs.nSubVols[0] * (iSub[1] + inArgs.nSubVols[1] * iSub[2]);
 		float* localCdf = &cdf[inArgs.nBins * idxSubVol]; // returns our part of array
 
-		for (int64_t iBin = 0; iBin < inArgs.nBins; iBin++) // reset bins
+		for (int iBin = 0; iBin < inArgs.nBins; iBin++) // reset bins
 			localCdf[iBin] = 0;
 
 		// calculate local maximum and minimum
@@ -230,13 +230,13 @@ __global__ void cdf_kernel(
 			startIdx[0] + inArgs.volSize[0] * (startIdx[1] + inArgs.volSize[1] * startIdx[2])];
 		float tempMax = firstVal; // temporary variable to reduce data access
 		float tempMin = firstVal;
-		for (int64_t iY = startIdx[2]; iY <= stopIdx[2]; iY++)
+		for (int iY = startIdx[2]; iY <= stopIdx[2]; iY++)
 		{
-			const int64_t yOffset = iY * inArgs.volSize[0] * inArgs.volSize[1];
-			for(int64_t iX = startIdx[1]; iX <= stopIdx[1]; iX++)
+			const int yOffset = iY * inArgs.volSize[0] * inArgs.volSize[1];
+			for(int iX = startIdx[1]; iX <= stopIdx[1]; iX++)
 			{
-				const int64_t xOffset = iX * inArgs.volSize[0];
-				for(int64_t iZ = startIdx[0]; iZ <= stopIdx[0]; iZ++)
+				const int xOffset = iX * inArgs.volSize[0];
+				for(int iZ = startIdx[0]; iZ <= stopIdx[0]; iZ++)
 				{
 					const float currVal = dataMatrix[iZ + xOffset + yOffset];
 					
@@ -261,19 +261,19 @@ __global__ void cdf_kernel(
 			1 : (tempMax - tempMin) / ((float) inArgs.nBins);
 
 		// sort values into bins which are above clipLimit
-		for (int64_t iY = startIdx[2]; iY <= stopIdx[2]; iY++)
+		for (int iY = startIdx[2]; iY <= stopIdx[2]; iY++)
 		{
-			const int64_t yOffset = iY * inArgs.volSize[0] * inArgs.volSize[1];
-			for(int64_t iX = startIdx[1]; iX <= stopIdx[1]; iX++)
+			const int yOffset = iY * inArgs.volSize[0] * inArgs.volSize[1];
+			for(int iX = startIdx[1]; iX <= stopIdx[1]; iX++)
 			{
-				const int64_t xOffset = iX * inArgs.volSize[0];
-				for(int64_t iZ = startIdx[0]; iZ <= stopIdx[0]; iZ++)
+				const int xOffset = iX * inArgs.volSize[0];
+				for(int iZ = startIdx[0]; iZ <= stopIdx[0]; iZ++)
 				{
 					const float currVal = dataMatrix[iZ + xOffset + yOffset]; 
 					// only add to histogram if above clip limit
 					if (currVal >= inArgs.noiseLevel)
 					{
-						int64_t iBin = (currVal - tempMin) / binRange;
+						int iBin = (currVal - tempMin) / binRange;
 
 						// special case for maximum values in subvolume (they gonna end up
 						// one index above)
@@ -291,7 +291,7 @@ __global__ void cdf_kernel(
 		// calculate cummulative sum and scale along y
 		localCdf[0] = 0; // we ignore the first bin since it is minimum anyway and should point to 0
 		float cdfTemp = 0;
-		for (int64_t iBin = 1; iBin < inArgs.nBins; iBin++)
+		for (int iBin = 1; iBin < inArgs.nBins; iBin++)
 		{
 			cdfTemp += localCdf[iBin];
 			localCdf[iBin] = cdfTemp;
@@ -301,14 +301,14 @@ __global__ void cdf_kernel(
 		const float cdfMax = localCdf[inArgs.nBins - 1];
 		if (cdfMax > 0)
 		{
-			for (int64_t iBin = 1; iBin < inArgs.nBins; iBin++)
+			for (int iBin = 1; iBin < inArgs.nBins; iBin++)
 			{
 				localCdf[iBin] = localCdf[iBin] / cdfMax;
 			}
 		}
 		else
 		{
-			for (int64_t iBin = 1; iBin < inArgs.nBins; iBin++)
+			for (int iBin = 1; iBin < inArgs.nBins; iBin++)
 			{
 				localCdf[iBin] = ((float) iBin) / ((float) inArgs.nBins);
 			}
@@ -323,19 +323,19 @@ class gridder
 {
 public:
 	// variables
-	int64_t volSize[3]; // size of full three dimensional volume [z, x, y]
-	int64_t nSubVols[3]; // number of subvolumes in zxy
-	int64_t sizeSubVols[3]; // size of each subvolume in zxy (should be uneven)
-	int64_t spacingSubVols[3]; // spacing of subvolumes (they can overlap)
-	int64_t origin[3]; // position of the very first element [0, 0, 0]
-	int64_t end[3]; // terminal value
-	int64_t nElements; // total number of elements
+	int volSize[3]; // size of full three dimensional volume [z, x, y]
+	int nSubVols[3]; // number of subvolumes in zxy
+	int sizeSubVols[3]; // size of each subvolume in zxy (should be uneven)
+	int spacingSubVols[3]; // spacing of subvolumes (they can overlap)
+	int origin[3]; // position of the very first element [0, 0, 0]
+	int end[3]; // terminal value
+	int nElements; // total number of elements
 	
 	// get functions
 	void calculate_nsubvols();
-	int64_t get_nSubVols(const uint8_t iDim) const {return nSubVols[iDim];};
-	int64_t get_nSubVols() const {return nSubVols[0] * nSubVols[1] * nSubVols[2];};
-	int64_t get_nElements() const;
+	int get_nSubVols(const uint8_t iDim) const {return nSubVols[iDim];};
+	int get_nSubVols() const {return nSubVols[0] * nSubVols[1] * nSubVols[2];};
+	int get_nElements() const;
 
 };
 
@@ -347,7 +347,7 @@ void gridder::calculate_nsubvols()
 	# pragma unroll
 	for (unsigned char iDim = 0; iDim < 3; iDim++)
 	{
-		const int64_t lastIdx = volSize[iDim] - 1;
+		const int lastIdx = volSize[iDim] - 1;
 		nSubVols[iDim] = (lastIdx - origin[iDim]) / spacingSubVols[iDim] + 1;
 		origin[iDim] = (sizeSubVols[iDim] - 1) / 2;
 		end[iDim] = origin[iDim] + (nSubVols[iDim] - 1) * spacingSubVols[iDim];
@@ -357,7 +357,7 @@ void gridder::calculate_nsubvols()
 }
 
 // returns number of element in the volume
-int64_t gridder::get_nElements() const
+int gridder::get_nElements() const
 {
 	return nElements;
 }
@@ -397,7 +397,7 @@ class histeq: public cudaTools, public gridder
 		float* minValBin; // maximum value occuring in each subvolume [iZ, iX, iY]
 		bool isMaxValBinAlloc = 0;
 		
-		int64_t nBins = 20; // number of histogram bins
+		int nBins = 20; // number of histogram bins
 		float noiseLevel = 0.1; // noise level threshold (clipLimit)
 
 		histeq();
@@ -448,7 +448,7 @@ void histeq::calculate_cdf_gpu()
 		inArgs.spacingSubVols[iDim] = spacingSubVols[iDim]; // size of each subvolume
 		inArgs.nSubVols[iDim] = nSubVols[iDim]; // number of subvolumes
 		inArgs.volSize[iDim] = volSize[iDim]; // overall size of data volume
-		inArgs.range[iDim] = (int64_t) (sizeSubVols[iDim] - 1) / 2; // range of each bin in each direction
+		inArgs.range[iDim] = (int) (sizeSubVols[iDim] - 1) / 2; // range of each bin in each direction
 		inArgs.origin[iDim] = origin[iDim];
 	}
 	inArgs.nBins = nBins;
@@ -458,7 +458,7 @@ void histeq::calculate_cdf_gpu()
 	float* cdf_dev; // cumulative distribution function [iBin, izSub, ixSub, iySub]
 	float* maxValBin_dev; // maximum value of current bin [izSub, ixSub, iySub]
 	float* minValBin_dev; // minimum value of current bin [izSub, ixSub, iySub]
-	const int64_t nCdf = nBins * get_nSubVols();
+	const int nCdf = nBins * get_nSubVols();
 
 	cudaError_t cErr;
 
@@ -547,8 +547,8 @@ void histeq::equalize_gpu()
 	float* cdf_dev; // array for cummulative distribution function [iBin, iZS, iXS, iYS]
 	float* minValBin_dev; // minimum value of each bin [iZS, iXS, iYS]
 	float* maxValBin_dev; // maximum value of each bin [iZS, iXS, iYS]
-	const int64_t nCdf = nBins * nSubVols[0] * nSubVols[1] * nSubVols[2];
-	const int64_t nSubs = nSubVols[0] * nSubVols[1] * nSubVols[2];
+	const int nCdf = nBins * nSubVols[0] * nSubVols[1] * nSubVols[2];
+	const int nSubs = nSubVols[0] * nSubVols[1] * nSubVols[2];
 
 	// allocate memory for main data array
 	cErr = cudaMalloc((void**)&dataMatrix_dev, nElements * sizeof(float));
@@ -627,9 +627,9 @@ void histeq::equalize_gpu()
 int main()
 {
 	// generate input volume matrix and assign random values to it
-	const int64_t volSize[3] = {600, 500, 400};
+	const int volSize[3] = {600, 500, 400};
 	float* inputVol = new float[volSize[0] * volSize[1] * volSize[2]];
-	for(int64_t iIdx = 0; iIdx < (volSize[0] * volSize[1] * volSize[2]); iIdx ++)
+	for(int iIdx = 0; iIdx < (volSize[0] * volSize[1] * volSize[2]); iIdx ++)
 		inputVol[iIdx] = ((float) rand()) / ((float) RAND_MAX);
 
 	// initialize some parameters
