+++ src/utest_binarization.cpp
@@ -18,15 +18,15 @@ int main()
 {
 
 	// define grid dimensions for testing
-	const vector3<int64_t> volSize(600, 500, 400);
+	const vector3<int> volSize(600, 500, 400);
 	const float clipLimit = 0.1;
-	const int64_t binSize = 250;
-	const vector3<int64_t> subVolSize(31, 31, 31);
-	const vector3<int64_t> subVolSpacing(20, 20, 20);
+	const int binSize = 250;
+	const vector3<int> subVolSize(31, 31, 31);
+	const vector3<int> subVolSpacing(20, 20, 20);
 
 	// generate input volume matrix and assign random values to it
 	float* inputVol = new float[volSize.elementMult()];
-	for(int64_t iIdx = 0; iIdx < volSize.elementMult(); iIdx ++)
+	for(int iIdx = 0; iIdx < volSize.elementMult(); iIdx ++)
 	{
 		inputVol[iIdx] = ((float) (iIdx % 2)) * 99.0 + 1.0;
 	}
@@ -62,7 +62,7 @@ int main()
 
 	// check if CDF is valid
 	// all bins until last one should have value 0.5 in CDF
-	for (int64_t iBin = 0; iBin < binSize; iBin++)
+	for (int iBin = 0; iBin < binSize; iBin++)
 	{
 		if (iBin < (binSize - 1))
 		{
@@ -88,9 +88,9 @@ int main()
 
 	// the output should now be all either 1s or 0s with an even distribution
 	float* outputVolCpu = histHandler.get_ptrOutput();
-	int64_t counterZero = 0;
-	int64_t counterOne = 0;
-	for (int64_t iElem = 0; iElem < (volSize.elementMult()); iElem++)
+	int counterZero = 0;
+	int counterOne = 0;
+	for (int iElem = 0; iElem < (volSize.elementMult()); iElem++)
 	{
 		if (outputVolCpu[iElem] == 0.0)
 		{
