+++ src/utest_cdf_gpu.cpp
@@ -20,19 +20,19 @@ int main(){
 	// define grid dimensions for testing
 	// initialize some parameters
 	const float clipLimit = 0.1;
-	const int64_t binSize = 10;
-	const vector3<int64_t> subVolSpacing = {20, 20, 20};
-	const vector3<int64_t> volSize = {600, 500, 400};
-	const vector3<int64_t> subVolSize(11, 11, 11);
+	const int binSize = 10;
+	const vector3<int> subVolSpacing = {20, 20, 20};
+	const vector3<int> volSize = {600, 500, 400};
+	const vector3<int> subVolSize(11, 11, 11);
 
 	// generate input volume matrix and assign random values to it
 	float* inputVol = new float[volSize[0] * volSize[1] * volSize[2]];
-	for(int64_t iIdx = 0; iIdx < (volSize[0] * volSize[1] * volSize[2]); iIdx ++)
+	for(int iIdx = 0; iIdx < (volSize[0] * volSize[1] * volSize[2]); iIdx ++)
 		inputVol[iIdx] = ((float) rand()) / ((float) RAND_MAX);
 		// this should generate a random number between 0 and 1
 
 
-	const int64_t iBin = rand() % binSize;
+	const int iBin = rand() % binSize;
 
 	histeq histHandler;
 	histHandler.set_nBins(binSize);
@@ -47,7 +47,7 @@ int main(){
 	
 	// backup the version of the CDF calculated with
 	float* cdf_bk = new float[histHandler.get_ncdf()];
-	for (int64_t iElem = 0; iElem < histHandler.get_ncdf(); iElem++)
+	for (int iElem = 0; iElem < histHandler.get_ncdf(); iElem++)
 	{
 		cdf_bk[iElem] = histHandler.get_cdf(iElem);
 	}
@@ -56,8 +56,8 @@ int main(){
 	// histogram calculation of CPU
 	histHandler.calculate_cdf();
 	bool isSame = 1;
-	int64_t countNotSame = 0;
-	for (int64_t iElem = 0; iElem < histHandler.get_ncdf(); iElem++)
+	int countNotSame = 0;
+	for (int iElem = 0; iElem < histHandler.get_ncdf(); iElem++)
 	{
 		const float deltaVal = abs(cdf_bk[iElem] - histHandler.get_cdf(iElem));
 		if (cdf_bk[iElem] != histHandler.get_cdf(iElem))
@@ -68,7 +68,7 @@ int main(){
 	}
 
 	printf("Displaying example CDF function:\n");
-	for (int64_t iBin = 0; iBin < binSize; iBin++)
+	for (int iBin = 0; iBin < binSize; iBin++)
 	{
 		printf("iBin: %d, GPU: %.3f, CPU: %.3f\n", 
 			(int) iBin, cdf_bk[iBin + 20], histHandler.get_cdf(iBin + 20));
