+++ src/utest_proc.cpp
@@ -20,17 +20,17 @@ int main()
 
 
 	// define testing parameters
-	const vector3<int64_t> volSize(600, 500, 300);
+	const vector3<int> volSize(600, 500, 300);
 	const float clipLimit = 0.01;
-	const int64_t binSize = 20;
-	const vector3<int64_t> subVolSize(31, 31, 31);
-	const vector3<int64_t> subVolSpacing(20, 20, 20);
+	const int binSize = 20;
+	const vector3<int> subVolSize(31, 31, 31);
+	const vector3<int> subVolSpacing(20, 20, 20);
 	
 	srand(1);
 
 	// generate input volume matrix and assign random values to it
 	float* inputVol = new float[volSize.x * volSize.y * volSize.z];
-	for(int64_t iIdx = 0; iIdx < (volSize.x * volSize.y * volSize.z); iIdx ++)
+	for(int iIdx = 0; iIdx < (volSize.x * volSize.y * volSize.z); iIdx ++)
 		inputVol[iIdx] = ((float) rand()) / ((float) RAND_MAX);
 		// this should generate a random number between 0 and 1
 
@@ -48,7 +48,7 @@ int main()
 
 	printf("Printing example bins\n");
 	float oldBinVal = 0;
-	for (int64_t iBin = 1; iBin < binSize; iBin++)
+	for (int iBin = 1; iBin < binSize; iBin++)
 	{
 		const float delta = histHandler.get_cdf(iBin) - oldBinVal;
 		printf("iBin = %d, Value = %.1f, Delta = %.4f\n", (int) iBin, histHandler.get_cdf(iBin), delta);
@@ -56,7 +56,7 @@ int main()
 	}
 
 	// first bin must be zero and last bin must be one
-	for (int64_t iSub = 0; iSub < histHandler.get_nSubVols(); iSub++)
+	for (int iSub = 0; iSub < histHandler.get_nSubVols(); iSub++)
 	{
 		if (histHandler.get_cdf(0, iSub) != 0.0)
 		{
