+++ src/utest_proc_gpu.cpp
@@ -18,15 +18,15 @@ int main()
 {
 
 	// define grid dimensions for testing
-	const vector3<int64_t> volSize(600, 500, 400);
+	const vector3<int> volSize(600, 500, 400);
 	const float clipLimit = 0.1;
-	const int64_t binSize = 20;
-	const vector3<int64_t> subVolSize(31, 31, 31);
-	const vector3<int64_t> subVolSpacing(20, 20, 20);
+	const int binSize = 20;
+	const vector3<int> subVolSize(31, 31, 31);
+	const vector3<int> subVolSpacing(20, 20, 20);
 
 	// generate input volume matrix and assign random values to it
 	float* inputVol = new float[volSize.elementMult()];
-	for(int64_t iIdx = 0; iIdx < (volSize.elementMult()); iIdx ++)
+	for(int iIdx = 0; iIdx < (volSize.elementMult()); iIdx ++)
 		inputVol[iIdx] = ((float) rand()) / ((float) RAND_MAX);
 		// this should generate a random number between 0 and 1
